Add files via upload
Experimental index.js and xviz-server.js files.
What I have done: 
Parsed all the x,y,z positions given by the /PathPlanner/desired_path ROS topic. These coordinates are to be used to generate the vertices for the polygon or polyline in XViz. I have set up what I believe could be one way of doing the XVizBuilder function for the path planner but I am unable to set up the properties for the polyline properly and the polyline properties show up as undefined ( as far as what the console log shows me).